Simplificando a seção HomeForm e utilizando contexto para os componentes dependentes
