+++ src/Sections/HomeForm/HomeForm.types.ts
@@ -2,5 +2,7 @@ export interface FormatCycle {
   task: string
   minutesOfDuration: number
   id: string
-  startData: Date
+  startDate: Date
+  interruptedDate?: Date
+  endDate?: Date
 }
